+++ CHANGELOG.md
@@ -7,6 +7,7 @@ ISSUES FIXED:
 * [51](https://github.com/perfectsense/gyro-azure-provider/issues/51): Implement Lifecycle for Azure Storage.
 * [55](https://github.com/perfectsense/gyro-azure-provider/issues/55): Azure VirtualMachine should Implement GyroInstance.
 * [56](https://github.com/perfectsense/gyro-azure-provider/issues/56): Implement Azure managed Identities.
+* [59](https://github.com/perfectsense/gyro-azure-provider/issues/59): Implement Role Assignment.
 * [62](https://github.com/perfectsense/gyro-azure-provider/issues/62): Add support to manage System Managed Service Identity for VM.
 * [64](https://github.com/perfectsense/gyro-azure-provider/issues/64): Expose ID field for CloudBlobContainerResource and CloudQueueResource.
 * [72](https://github.com/perfectsense/gyro-azure-provider/issues/72): Expose object id for System Managed Identity on Virtual Machine and Scale Set resources.